requirements.txt
POT
numpy
scipy
matplotlib
gurobi